refactor: move paas package to internal/cli/clients
because now it only has the 2 carrier "clients". Further refactoring
will deprecate these packages by splitting thier methods to more
appropriate places. E.g. anything that has to do with applications
belongs to the application package. Same for services etc.
