+++ internal/cli/disable.go
@@ -4,9 +4,9 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/suse/carrier/deployments"
+	"github.com/suse/carrier/internal/cli/clients"
 	"github.com/suse/carrier/internal/duration"
 	"github.com/suse/carrier/kubernetes"
-	"github.com/suse/carrier/paas"
 	"github.com/suse/carrier/termui"
 )
 
@@ -59,7 +59,7 @@ func DisableGoogle(cmd *cobra.Command, args []string) error {
 
 func UninstallDeployment(cmd *cobra.Command, deployment kubernetes.Deployment, successMessage string) error {
 	uiUI := termui.NewUI()
-	installClient, installCleanup, err := paas.NewInstallClient(cmd.Flags(), &kubernetes.InstallationOptions{})
+	installClient, installCleanup, err := clients.NewInstallClient(cmd.Flags(), &kubernetes.InstallationOptions{})
 	defer func() {
 		if installCleanup != nil {
 			installCleanup()
