+++ internal/cli/enable.go
@@ -4,9 +4,9 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/suse/carrier/deployments"
+	"github.com/suse/carrier/internal/cli/clients"
 	"github.com/suse/carrier/internal/duration"
 	"github.com/suse/carrier/kubernetes"
-	"github.com/suse/carrier/paas"
 	"github.com/suse/carrier/termui"
 )
 
@@ -76,7 +76,7 @@ func EnableGoogle(cmd *cobra.Command, args []string) error {
 
 func InstallDeployment(cmd *cobra.Command, deployment kubernetes.Deployment, opts kubernetes.InstallationOptions, successMessage string) error {
 	uiUI := termui.NewUI()
-	installClient, installCleanup, err := paas.NewInstallClient(cmd.Flags(), &opts)
+	installClient, installCleanup, err := clients.NewInstallClient(cmd.Flags(), &opts)
 	defer func() {
 		if installCleanup != nil {
 			installCleanup()
